Refactored types #422
